simplified the Contract hierarchy.
