--- src/main/scala/org/economicsl/auctions/OrderLike.scala
@@ -1,39 +1,0 @@
-/*
-Copyright (c) 2017 KAPSARC
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-   http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-package org.economicsl.auctions
-
-import org.economicsl.core.Tradable
-
-
-/** Mixin trait indicating that a `Contract` is `OrderLke`.
-  *
-  * @author davidrpugh
-  * @since 0.1.0
-  */
-trait OrderLike[+T <: Tradable] {
-  this: Contract =>
-
-  /** The type of tradable for which the order has been issued. */
-  def tradable: T
-
-}
-
-
-
-
-
-
-
