+++ src/main/scala/org/economicsl/auctions/ReferenceGenerator.scala
@@ -1,0 +1,16 @@
+package org.economicsl.auctions
+
+
+import org.economicsl.core.util.UUIDGenerator
+
+
+trait ReferenceGenerator
+    extends UUIDGenerator {
+
+  protected def randomReference(): Reference = {
+    random()
+  }
+
+}
+
+
