+++ src/main/scala/org/economicsl/auctions/TokenGenerator.scala
@@ -1,0 +1,13 @@
+package org.economicsl.auctions
+
+import org.economicsl.core.util.UUIDGenerator
+
+
+trait TokenGenerator
+  extends UUIDGenerator {
+
+  protected def randomToken(): Token = {
+    random()
+  }
+
+}
